Health Bar Pog

+++ 1610Fall2020 ass0/Assets/ScripleObjects/FloatData.cs
@@ -1,14 +1,13 @@
 ﻿using UnityEngine;
 
 [CreateAssetMenu]
-public class IntData : ScriptableObject
+public class FloatData : ScriptableObject
 {
-    public int value;
+    public float value;
 
-    public void UpdateValue(int number)
+    public void UpdateValue(float number)
     {
         value += number;
         Debug.Log(value + " " + this.name);
     }
-    
 }
